fixed division by 0

+++ src/evaluator.h
@@ -8,7 +8,7 @@ void infix_to_postfix(Token infix[EQUATION_LEN + 1], Token postfix[EQUATION_LEN 
 
 int evaluate_expression(Token tokens[EQUATION_LEN + 1]);
 
-int evaluate_string(const char *buf, int len, double *result);
+int evaluate_string(const char *buf, int len, int *result);
 
 int process_line(char *line, int print_message);
